requirements.txt
boto==2.49.0
boto3==1.13.8
botocore==1.16.8
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
docutils==0.15.2
gensim==3.8.3
idna==2.9
jmespath==0.10.0
joblib==0.14.1
kiwisolver==1.2.0
mahotas==1.4.10
matplotlib==3.2.1
nltk==3.5
numpy==1.18.4
pandas==1.0.3
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.5.7
requests==2.23.0
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
smart-open==2.0.0
tqdm==4.46.0
urllib3==1.25.9
